refactor: move large-module tests to owned seams
Keep the shared secure account-session implementation intact while moving its inline tests beside cloud.rs. Do the same for bundled system-skill migration tests, so production modules own behavior and tests through explicit seams without making test bodies part of the production source budget.

Owned production Rust falls from 648531 to 647385 lines (net -1146). cloud.rs is 991 lines and skills/system.rs is 515; their unchanged test suites are 652 and 489 lines. No source-structure budget value changes.

Verified with the full 200-test CLI library suite, all 21 bundled-system-skill tests, strict CLI clippy, strict TUI all-target clippy modulo four pre-existing unrelated lints, CLI/TUI all-target checks, formatting, source-structure budget, and diff checks.

(cherry picked from commit e0918d132f4bc28121ad556e0125cb622db6a62d)
